% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{ARX}
\alias{ARX}
\title{Automatic train an Autoregressive model}
\usage{
ARX(formula, ...)
}
\arguments{
\item{formula}{Model specification (currently not in use).}

\item{...}{Further arguments passed to \code{auto_arx()}.}
}
\value{
An object of class \code{ARX}.
}
\description{
This function trains an Autoregressive model (ARX)
  to a univariate time series.
}
