% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{auto_arx}
\alias{auto_arx}
\title{Automatic train an Autoregressive model}
\usage{
auto_arx(
  .data,
  specials,
  lags = NULL,
  fourier = NULL,
  xreg = NULL,
  dy = 0,
  dx = 0,
  inf_crit = "aic",
  n_models = 500,
  n_seed = 42,
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{.data}{A \code{tsibble} containing the time series data.}

\item{specials}{Currently not is use.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{inf_crit}{Character value. The information criterion used for variable selection \code{inf_crit = c("aic", "aicc", "bic")}.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
An object of class \code{ARX}.
}
\description{
This function trains an Autoregressive model (ARX)
  to a univariate time series.
}
