% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_esn.R
\name{auto_esn}
\alias{auto_esn}
\title{auto_esn}
\usage{
auto_esn(
  data,
  n_res = 100,
  n_sample = 200,
  max_lag = 6,
  max_trig = c(3, 3),
  n_diff = 1
)
}
\arguments{
\item{data}{A tsibble containing the time series data. Must have column "time" with time index (date or date-time).}

\item{n_sample}{Integer value. The number of samples for the random grid.}

\item{max_lag}{Integer value. The maximum number of non-seasonal lags.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{max_cycle}{Integer vector. The maximum number of trigonometric terms per season.}

\item{diff}{Logical value. If \code{TRUE}, the time series is modeled in first difference.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}
}
\value{
An object of class \code{ESN}.
}
\description{
auto_esn ...
}
