% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_const}
\alias{create_const}
\title{Create constant (intercept term).}
\usage{
create_const(n_obs)
}
\arguments{
\item{n_obs}{Integer value. The number of observations.}
}
\value{
y_const Numeric matrix with dimension (n_obs x 1).
}
\description{
This functions creates the constant (intercept term) for the design matrix as numeric matrix.
}
