% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_fourier}
\alias{create_fourier}
\title{Create fourier terms.}
\usage{
create_fourier(times, n_terms, period)
}
\arguments{
\item{times}{Integer vector. Regular sequence with integers.}

\item{n_terms}{Integer vector. The number of fourier terms per period (i.e. the number of sines and cosines for each period).}

\item{period}{Integer vector. The periodicity of the time series.}
}
\value{

}
\description{
This function creates the fourier terms for the design matrix as numeric matrix.
}
