% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_grid_fourier}
\alias{create_grid_fourier}
\title{Create a grid of fourier terms for best subset regression}
\usage{
create_grid_fourier(n_fourier, period)
}
\arguments{
\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles) per period.}

\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data \code{period = c(12)}, for hourly data \code{period = c(24, 168)}).}
}
\value{
blocks A tibble containing zeros and ones for all combinations of fourier terms.
}
\description{
\code{create_grid_fourier} creates a grid of all combinations
   of fourier terms. One fourier term is always a pair of sine and cosine
   terms. If higher order fourier terms are used, it is assumed that the
   previous terms are added too.
}
