% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_lags}
\alias{create_lags}
\title{Create lagged variables of a matrix}
\usage{
create_lags(data, lags)
}
\arguments{
\item{data}{Numeric vector or matrix. Each column is a variable and each row an observation.}

\item{lags}{List containing vectors with the number of lags (in units of observations) per variable.}
}
\value{
y_lag Numeric matrix with lagged variables of the input data.
}
\description{
Create lagged variables of a matrix, shifting each column
  back by a given number of observations.
}
