% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_revolved.R
\name{create_revolved}
\alias{create_revolved}
\title{Create lagged variables of a matrix for iterative forecasting}
\usage{
create_revolved(data, lags, n_ahead)
}
\arguments{
\item{data}{A numeric matrix with lagged variables for forecasting.}

\item{lags}{A list containing vectors with the number of lags (in units of observations) per variable.}

\item{n_ahead}{Forecast horizon (n-step ahead).}
}
\value{
y_lag A numeric matrix with the lagged variables of the input data for iterative forecasting.
}
\description{
Create lagged variables of a matrix for iterative forecasting, shifting each column back by a given number of observations and fill with NAs for the updates.
}
