% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_season}
\alias{create_season}
\title{Create trigonometric terms.}
\usage{
create_season(times, k, period)
}
\arguments{
\item{times}{Integer value. The number of observations.}

\item{k}{Integer vector. The number of seasonal cycles per period.}

\item{period}{Integer vector. The periodicity of the time series.}
}
\value{

}
\description{
This function creates the trigonometric terms for the design matrix as numeric matrix.
}
