% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_spec}
\alias{create_spec}
\title{Create model specification}
\usage{
create_spec(model_layers)
}
\arguments{
\item{model_layers}{List containing the number of inputs (n_inputs), reservoir size (n_res) and the number of outputs (n_outputs).}
}
\value{
model_spec Character value. The model specification as string.
}
\description{
This function creates the model specification (short summary) as a string.
}
