% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_trig}
\alias{create_trig}
\title{Create trigonometric terms.}
\usage{
create_trig(times, n_trig, period)
}
\arguments{
\item{times}{Integer value. The number of observations.}

\item{n_trig}{Integer vector. The number of seasonal cycles per period (i.e. the number of sines and cosines for each frequency).}

\item{period}{Integer vector. The periodicity of the time series.}
}
\value{

}
\description{
This function creates the trigonometric terms for the design matrix as numeric matrix.
}
