% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_win.R
\name{create_win}
\alias{create_win}
\title{Input weight matrix}
\usage{
create_win(n_inputs, n_res, scale_runif)
}
\arguments{
\item{n_inputs}{Number of input features}

\item{n_res}{Number of internal states within the reservoir}

\item{scale_runif}{Lower and upper bound of the uniform distribution}
}
\value{
win A list containing the input weight matrices.
}
\description{
This function creates the random input weight matrices
}
