% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wres.R
\name{create_wres}
\alias{create_wres}
\title{Reservoir weights}
\usage{
create_wres(n_res, rho, density, scale_runif, symmetric = FALSE)
}
\arguments{
\item{n_res}{Number of internal states within the reservoir}

\item{density}{The parameter defines the connectivity of the reservoir weight matrix (dense or sparse)}

\item{unif_range}{Lower and upper bound of the uniform distribution}
}
\value{
wres The final reservoir weight matrix
}
\description{
This function creates the random reservoir weight matrix (scaled to spectral radius rho)
}
