% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_data}
\alias{diff_data}
\title{Calculate (non-seasonal) differences of a numeric matrix}
\usage{
diff_data(data, n_diff)
}
\arguments{
\item{data}{Numeric matrix.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}
}
\value{
y_diff Numeric matrix with the differenced data.
}
\description{
This function takes a numeric matrix and calculates 
   (non-seasonal) differences for each column.
}
