% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_num}
\alias{diff_num}
\title{Calculate seasonal and non-seasonal differences of a numeric vector.}
\usage{
diff_num(y, period, n_sdiff, n_diff)
}
\arguments{
\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer value. The number of seasonal differences.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}

\item{data}{Numeric vector.}
}
\value{
y_diff Numeric vector with the differenced data.
}
\description{
This function takes a numeric vector and calculates seasonal and non-seasonal differences.
}
