% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_center}
\alias{estimate_center}
\title{Estimate the central tendency.}
\usage{
estimate_center(x, type = "median", ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{type}{Character value. Possible values are \code{"mean"}, \code{"median"} or \code{"mode"}.}

\item{...}{Further arguments passed to \code{base::mean()}, \code{stats::median()} or \code{stats::densitiy()}.}
}
\value{
out The central tendency as numeric value.
}
\description{
The function estimates the central tendency either by the
   arithmetic mean, median or mode of a numeric vector.
}
