% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glmnet.R
\name{estimate_dof}
\alias{estimate_dof}
\title{Estimate effective degrees of freedom}
\usage{
estimate_dof(X, lambda)
}
\arguments{
\item{X}{Numeric matrix. The design matrix containing the predictor variables.}

\item{lambda}{Numeric value. The regularization parameter.}
}
\value{
Numeric value.
}
\description{
The function estimates the effective degrees of freedom.
}
