% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimate_mode}
\alias{estimate_mode}
\title{Calculates the mode of a distribution based on Kernel Density Estimation}
\usage{
estimate_mode(x, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{...}{Further arguments passed to \code{stats::densitiy()}.}
}
\value{
mode Numeric value. The mode of the distribution.
}
\description{
Calculates the mode of a distribution based on Kernel Density Estimation.
}
