% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{extract_arx}
\alias{extract_arx}
\title{Extract and return values from a trained ARX as tibble}
\usage{
extract_arx(object)
}
\arguments{
\item{object}{An object of class \code{mdl_df}.}
}
\value{
A tibble containing the hyper-parameters.
}
\description{
Extract and return values from a trained ARX
  as tibble. The function works only if the
  model within the \code{mdl_df} is of class \code{ARX}.
  The extracted values are are stored in a tibble with
  the following columns:
  
  \itemize{
    \item{\code{spec}: Character value. Succinct summary of model specifications.}
    \item{\code{n_inputs}: Integer value. The number of model inputs.}
    \item{\code{n_outputs}: Integer value. The number of model outputs.}
    \item{\code{lags}: A \code{list} containing integer vectors with the lags associated with each input variable.}
    \item{\code{fourier}: A \code{list} containing the fourier terms.}
    \item{\code{dy}: Integer vector. The nth-differences of the response variable.}
    \item{\code{dx}: Integer vector. The nth-differences of the exogenous variable.}
    }
}
