% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{fbl_train_esn}
\alias{fbl_train_esn}
\title{Train an Echo State Network (ESN)}
\usage{
fbl_train_esn(
  .data,
  specials,
  const = TRUE,
  lags = NULL,
  n_fourier = NULL,
  n_initial = 10,
  n_res = 200,
  n_seed = 42,
  n_sample = 5000,
  alpha = seq(0, 1, 0.25),
  rho = seq(0.25, 2.5, 0.25),
  lambda = c(1, 5, 10),
  density = 0.1,
  scale_inputs = c(-1, 1),
  scale_runif = c(-0.5, 0.5),
  inf_crit = "bic",
  type = "mode",
  ...
)
}
\arguments{
\item{.data}{Input data as \code{tsibble}.}

\item{specials}{Specials as list defined in \code{specials_esn}.}

\item{...}{Further arguments passed to \code{train_esn()}.}

\item{max_lag}{Integer value. Maximum number of non-seasonal lags.}
}
\value{
An object of class \code{ESN}.
}
\description{
Train an Echo State Network (ESN).
}
