% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ridge.R
\name{fit_ridge}
\alias{fit_ridge}
\title{Estimate a linear model via ridge regression}
\usage{
fit_ridge(x, y, lambda)
}
\arguments{
\item{y}{Numeric matrix. The response variable(s).}

\item{X}{Numeric matrix. The design matrix containing the predictor variables.}
}
\value{
A list containing the estimated coefficients, fitted values and some
  goodness-of-fit metrics.
}
\description{
Estimate a linear model via Ordinary Least Squares (OLS). 
  \code{fit_lm()} is a wrapper function for \code{stats::lm.fit()} with some
  additional output like goodness-of-fit metrics (e.g. information criteria). 
  The function is slightly faster than \code{stats::lm()}, bit most 
  importantly, the resulting object size is much smaller.
}
