% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rlm.R
\name{fit_rlm}
\alias{fit_rlm}
\title{Estimate a linear model via robust regression}
\usage{
fit_rlm(x, y)
}
\arguments{
\item{y}{Numeric matrix. The response variable(s).}

\item{X}{Numeric matrix. The design matrix containing the predictor variables.}
}
\value{
A list containing the estimated coefficients, fitted values and some
  goodness-of-fit metrics.
}
\description{
Estimate a linear model via robust regression. 
  \code{fit_rlm()} is a wrapper function for \code{MASS::rlm()} with some
  additional output like goodness-of-fit metrics (e.g. information criteria). 
  The resulting object size is much smaller.
}
