% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{fitted.ARX}
\alias{fitted.ARX}
\title{Extract fitted values from a trained ARX}
\usage{
\method{fitted}{ARX}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ARX}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values extracted from the object.
}
\description{
Extract fitted values from a trained ARX.
}
