% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{fitted.ESN}
\alias{fitted.ESN}
\title{Extract fitted values from a trained ESN}
\usage{
\method{fitted}{ESN}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ESN}.}

\item{...}{Currently not in use.}
}
\value{
Fitted values extracted from the object.
}
\description{
Extract fitted values from a trained ESN.
}
\examples{
library(tsibble)
library(fable)
AirPassengers \%>\%
as_tsibble() \%>\%
model("ESN" = ESN(value)) \%>\%
fitted()

}
