% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{fitted.fbl_esn}
\alias{fitted.fbl_esn}
\title{Extract fitted values from a ELM.}
\usage{
\method{fitted}{fbl_esn}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Additional arguments passed to ...}
}
\value{

}
\description{
Extract fitted values from a ELM.
}
