% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{forecast.ARX}
\alias{forecast.ARX}
\title{Forecast a trained ARX}
\usage{
\method{forecast}{ARX}(object, new_data, specials = NULL, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ARX}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Currently not in use}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained ARX.
}
