% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{forecast.ESN}
\alias{forecast.ESN}
\title{Forecast a trained ESN}
\usage{
\method{forecast}{ESN}(object, new_data, specials = NULL, n_sim = 100, ...)
}
\arguments{
\item{object}{Trained model}

\item{new_data}{Forecast horizon}

\item{specials}{NULL}

\item{...}{Further arguments passed toi bla}
}
\value{
A fable
}
\description{
Forecast a trained ESN
}
