% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{forecast.fbl_esn}
\alias{forecast.fbl_esn}
\title{Forecast a ELM.}
\usage{
\method{forecast}{fbl_esn}(object, new_data, specials = NULL, times = 1000, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecasts.}

\item{new_data}{A tsibble containing future information used to forecast.}

\item{specials}{Specials are currently not in use for MLPs.}

\item{times}{The number of sample paths to use in estimating the forecast distribution when \code{bootstrap = TRUE}.}

\item{...}{Additional arguments for forecast model method.}
}
\value{
An object of class "fable".
}
\description{
Forecast a Extreme Learning Machine (ELM).
}
