% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_arx.R
\name{forecast_arx}
\alias{forecast_arx}
\title{Forecast a trained Autoregressive model}
\usage{
forecast_arx(object, n_ahead = 12, xreg = NULL)
}
\arguments{
\item{object}{An object of class \code{ARX}. The result of a call to \code{train_arx()} or \code{auto_arx()}.}

\item{n_ahead}{Integer value. The number of periods for forecasting (forecast horizon).}

\item{xreg}{A \code{tsibble} containing exogenous variables.}
}
\value{
A \code{list} containing:
   \itemize{
      \item{\code{point}: Numeric vector containing the point forecasts.}
      \item{\code{method}: A \code{list} containing several objects and meta information of the trained ESN (weight matrices, hyperparameters, model metrics, etc.).}
      \item{\code{n_ahead}: Integer value. The number of periods for forecasting (forecast horizon).}
      }
}
\description{
Forecast a trained Autoregressive model.
}
