% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{glance.ARX}
\alias{glance.ARX}
\title{Summary of model fit}
\usage{
\method{glance}{ARX}(object)
}
\arguments{
\item{object}{An object of class \code{ARX}.}
}
\value{
Summary statistics extracted from the object.
}
\description{
Return summary statistics from a trained ARX as tibble.
 \itemize{
   \item{\code{df}: Effective degrees of freedom.}
   \item{\code{aic}: Akaike information criterion.}
   \item{\code{bic}: Bayesian information criterion.}
   \item{\code{hqc}: Hannan-Quinn information criterion.}
      }
}
