% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_diff_num}
\alias{inv_diff_num}
\title{Integrate seasonal and non-seasonal differences of a numeric vector ('inverse differencing').}
\usage{
inv_diff_num(y, y_diff, period, n_sdiff, n_diff, type)
}
\arguments{
\item{y}{Numeric vector containing the original data.}

\item{y_diff}{Numeric vector containing the differenced data.}

\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer value. The number of seasonal differences.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}

\item{type}{Character value. Either \code{type = "inner"} or \code{type = "outer"}. This distinction is relevant for the starting values of the integration.}
}
\value{
y_int Numeric vector with the inverse differenced data.
}
\description{
This function takes a numeric vector and integrates seasonal and non-seasonal differences ('inverse differencing').
}
