% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_diff_vec}
\alias{inv_diff_vec}
\title{Integrate (non-seasonal) differences of a numeric vector
   ("inverse difference")}
\usage{
inv_diff_vec(y, y_diff, n_diff)
}
\arguments{
\item{y}{Numeric vector containing the original data.}

\item{y_diff}{Numeric vector containing the differenced data.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}
}
\value{
y_int Numeric vector with the inverse differenced data.
}
\description{
This function takes a numeric vector and integrates
   (non-seasonal) differences ("inverse difference").
}
