% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_trans_data.R
\name{inv_trans_data}
\alias{inv_trans_data}
\title{Inverse transforms the columns of a numeric matrix}
\usage{
inv_trans_data(object, new_data)
}
\arguments{
\item{object}{A list with objects of class orderNorm.}

\item{new_data}{New data for inverse transforming (e.g. forecast)}
}
\value{
data Inverse transformed data as numeric matrix.
}
\description{
Inverse transforms the columns of a numeric matrix using inverse ordered quantile normalizing transformation (OQR)
}
