% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{is.esn}
\alias{is.esn}
\title{Checks if object is of class "esn"}
\usage{
is.esn(object)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Returns \code{TRUE} if the object is of class "esn".
}
