% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_data}
\alias{m4_data}
\title{Deprecated name for \code{m4_monthly_subset}}
\format{
See \code{\link{m4_monthly_subset}}.
}
\usage{
data(m4_data)
}
\description{
\code{m4_data} has been renamed to \code{m4_monthly_subset}. The old name is
retained for backward compatibility with earlier versions of \code{echos},
but users should prefer \code{m4_monthly_subset} in new code.
}
\examples{
data(m4_data)
head(m4_data)
}
\seealso{
\code{\link{m4_monthly_subset}}
}
\keyword{datasets}
