% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{model_sum.ARX}
\alias{model_sum.ARX}
\title{Provide a succinct summary of a trained ARX}
\usage{
\method{model_sum}{ARX}(object)
}
\arguments{
\item{object}{An object of class \code{ARX}.}
}
\value{
Model summary extracted from the object.
}
\description{
Provide a succinct summary of a trained ARX.
}
