% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{model_sum.fbl_esn}
\alias{model_sum.fbl_esn}
\title{Provide a succinct summary of the MLP.}
\usage{
\method{model_sum}{fbl_esn}(x)
}
\arguments{
\item{object}{The MLP to summarize.}
}
\value{

}
\description{
Provide a succinct summary of the MLP.
}
