% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{pars}
\alias{pars}
\title{Return hyper-parameters from a trained ESN as tibble}
\usage{
pars(object, ...)
}
\arguments{
\item{object}{An object of class \code{mdl_df}.}
}
\value{
A tibble containing the hyper-parameters.
}
\description{
Return the hyper-parameters from a
  trained ESN as tibble. The function works only if the
  model within the \code{mdl_df} is of class \code{ESN}.
}
