% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{plot.esn}
\alias{plot.esn}
\title{Plot actuals, fitted and residuals of a trained ESN model.}
\usage{
\method{plot}{esn}(object)
}
\arguments{
\item{object}{An object of class \code{esn}.}
}
\value{
Standard plots.
}
\description{
Provide standard plots of a trained ESN model.
}
