% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{plot.forecast_esn}
\alias{plot.forecast_esn}
\title{Plot point forecasts and actuals of a trained ESN model.}
\usage{
\method{plot}{forecast_esn}(object, fitted = FALSE, test = NULL)
}
\arguments{
\item{object}{An object of class \code{forecast_esn}.}
}
\value{
Line chart of point forecast and actual values.
}
\description{
Plot point forecasts and actuals of a trained ESN model as line
  chart. Optionally, fitted values and test data can be added to the plot.
}
