% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot the forecast of an Echo State Network.}
\usage{
plot_forecast(
  object,
  include = NULL,
  add_test = NULL,
  add_fitted = FALSE,
  add_paths = FALSE,
  add_legend = TRUE,
  legend_position = "right",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  xlab = "Time",
  ylab = NULL,
  color_fcst = "#0000AA",
  color_fitted = "#0000AA",
  color_actual = "black",
  color_test = "#D55E00",
  color_paths = "grey35",
  size_fcst = 1,
  size_fitted = 0.5,
  size_actual = 0.5,
  size_test = 1,
  size_paths = 0.1
)
}
\arguments{
\item{object}{An object of class "forecast". The result of a call to forecast_esn(...).}

\item{include}{Integer value. Number of actual values (and fitted values) to be included in the graph.}

\item{add_test}{A tsibble containing the test data (hold-out sample).}

\item{add_fitted}{Logical value. If TRUE, fitted values are plotted along the actuals.}

\item{add_paths}{Logical value. If TRUE, simulations are plotted.}

\item{add_legend}{Logical value. If TRUE, a legend is added.}

\item{legend_position}{Character value. Position of the legend ("right", "left", "top", "bottom").}

\item{title}{Title for the plot.}

\item{subtitle}{Subtitle for the plot.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{color_fcst}{Line color of the point forecasts.}

\item{color_fitted}{Line color of the fitted values.}

\item{color_actual}{Line color of the actual values.}

\item{color_test}{Line color of test data.}

\item{color_paths}{Line color of the simulated future sample paths.}

\item{size_fcst}{Line width of point forecast.}

\item{size_fitted}{Line width of fitted values.}

\item{size_actual}{Lines width of actual values.}

\item{size_test}{Line width of test data.}

\item{size_paths}{Line width of simulated future sample paths.}
}
\value{
p An object of class ggplot.
}
\description{
Autoplot the forecasts from an Echo State Network (univariate and multivariate models).
}
