% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{predict_esn}
\alias{predict_esn}
\title{Forecast a fitted ESN.}
\usage{
predict_esn(
  win,
  wres,
  wout,
  n_ahead,
  alpha,
  lags,
  inputs,
  states_train,
  innov = NULL
)
}
\arguments{
\item{win}{Numeric matrix. Weights for the input variables.}

\item{wres}{Numeric matrix. Weights for the reservoir.}

\item{wout}{Numeric matrix. Weights for output variables (estimated coefficients from ridge regression).}

\item{n_ahead}{Integer value. The forecast horizon (n-step ahead).}

\item{alpha}{Numeric value. The Leakage rate (smoothing parameter).}

\item{lags}{List containing integer vectors with the lags associated with each output variable.}

\item{inputs}{Numeric matrix. Initialized input features (gets updated during forecasting process).}

\item{states_train}{Numeric matrix. Internal states from training (necessary due to last values).}

\item{innov}{Numeric matrix. The innovations for simulation.}
}
\value{
fcst Numeric matrix with point forecasts.

states_fcst Numeric matrix with internal states used for forecasting.
}
\description{
Calculate point forecasts for a fitted ESN (internally).
}
