% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{print.esn}
\alias{print.esn}
\title{Print model specification of the trained ESN model}
\usage{
\method{print}{esn}(x, ...)
}
\arguments{
\item{x}{An object of class \code{esn}.}

\item{...}{Currently not in use.}
}
\value{
Print specification of the trained ESN model.
}
\description{
Print model specification of the trained ESN model.
}
\examples{
xdata <- as.numeric(AirPassengers)
xmodel <- train_esn(y = xdata)
print(xmodel)

}
