% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_const}
\alias{random_const}
\title{Create random grid of constant terms}
\usage{
random_const(y_const, n_sample = 5000)
}
\arguments{
\item{y_const}{Numeric matrix. The result of a call to \code{create_const()}.}

\item{n_sample}{Integer value. The number of random samples.}
}
\value{
out Random grid of constant terms as \code{tibble}.
}
\description{
The function creates a \code{tibble} with random combinations
   of constant terms (sampling with replacement). The number of combinations (rows)
   is given by \code{n_sample} and the number of columns is always one.
}
