% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_fourier}
\alias{random_fourier}
\title{Create random grid of fourier terms}
\usage{
random_fourier(n_fourier, period, n_sample = 5000)
}
\arguments{
\item{n_fourier}{Integer vector. The number of fourier terms per seasonal period.}

\item{period}{Integer vector. The seasonal periods.}

\item{n_sample}{Integer value. The number of random samples.}
}
\value{
out Random grid of fourier terms as \code{tibble}.
}
\description{
The function creates a \code{tibble} with random combinations
   of fourier terms (sampling with replacement). The number of combinations
   (rows) is given by \code{n_sample} and the number of columns is
   determined by \code{n_fourier} and \code{period}.
}
