% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_lags}
\alias{random_lags}
\title{Create random grid of lags}
\usage{
random_lags(y_lag, n_sample = 1000)
}
\arguments{
\item{y_lag}{Numeric matrix. The result of a call to \code{create_lags()}.}

\item{n_sample}{Integer value. The number of random samples.}
}
\value{
out Random grid of lags as \code{tibble}.
}
\description{
The function creates a \code{tibble} with random combinations
   of lags (sampling with replacement). The number of combinations (rows)
   is given by \code{n_sample} and the number of columns is determined by
   \code{y_lag}.
}
