% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{report.ARX}
\alias{report.ARX}
\title{Provide a detailed summary of a trained ARX}
\usage{
\method{report}{ARX}(object)
}
\arguments{
\item{object}{An object of class \code{ARX}.}
}
\value{
Print detailed model summary.
}
\description{
Provide a detailed summary of a trained ARX.
}
