% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{report.ESN}
\alias{report.ESN}
\title{Provide a detailed summary of a trained ESN}
\usage{
\method{report}{ESN}(object)
}
\arguments{
\item{object}{An object of class \code{ESN}.}
}
\value{

}
\description{
Provide a detailed summary of a trained ESN.
}
