% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_data.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale (inverse scaling) the columns of a numeric matrix}
\usage{
rescale_data(data, old_range, new_range)
}
\arguments{
\item{data}{is a vector or matrix containing the values to be rescaled. Each column is a variable and each row an observation.}

\item{old_range}{A numeric matrix with range (min and max) of original data.}

\item{new_range}{A numeric vector with new (scaled) interval.}
}
\value{
data A numeric matrix with rescaled columns.
}
\description{
Recale (inverse scaling) the columns of a numeric matrix by applying the transformation backwards to original range.
}
