% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{reservoir}
\alias{reservoir}
\title{Return the reservoir from a trained ESN as tibble}
\usage{
reservoir(object)
}
\arguments{
\item{object}{An object of class \code{mdl_df}.}
}
\value{
A tibble containing the reservoir (internal states).
}
\description{
Return the reservoir (internal states) from a
  trained ESN as tibble. The function works only if the
  model within the \code{mdl_df} is of class \code{ESN}.
}
