% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{residuals.ARX}
\alias{residuals.ARX}
\title{Extract residuals from a trained ARX}
\usage{
\method{residuals}{ARX}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ARX}.}

\item{...}{Currently not in use.}
}
\value{
Residuals extracted from the object.
}
\description{
Extract residuals from a trained ARX.
}
