% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{residuals.fbl_esn}
\alias{residuals.fbl_esn}
\title{Extract residuals from a MLP.}
\usage{
\method{residuals}{fbl_esn}(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Additional arguments passed to ...}
}
\value{

}
\description{
Extract residuals from a MLP.
}
