% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rotate}
\alias{rotate}
\title{Rotate a matrix}
\usage{
rotate(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix.
}
\description{
Little helper function to rotate a matrix. If you apply rotate
   two times, the matrix is flipped (\code{rotate(rotate(x))}).
}
