% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_reservoir}
\alias{run_reservoir}
\title{Run reservoir}
\arguments{
\item{input}{Numeric matrix containing the input features}

\item{win}{Numeric matrix. The input weight matrix.}

\item{wres}{Numeric matrix. The reservoir weight matrix.}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter).}
}
\value{
states train Numeric matrix with the internal states.
}
\description{
Run reservoir creates the internal states for the ESN.
}
