% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_data.R
\name{scale_data}
\alias{scale_data}
\title{Scale the columns of a numeric matrix}
\usage{
scale_data(data, new_range = c(-1, 1))
}
\arguments{
\item{data}{A numeric matrix containing the values to be scaled. Each column is a variable and each row an observation.}

\item{new_range}{The range for scaling (first value represents the replacement for the min value, the second is the substitute for the max value, default c(-1, 1)).}
}
\value{
data A numeric matrix with scaled columns.
}
\description{
Scale the columns of a numeric matrix to a specific interval
}
