% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_inputs}
\alias{select_inputs}
\title{Model selection}
\usage{
select_inputs(
  data,
  lags,
  n_fourier,
  period,
  const,
  n_sdiff,
  n_diff,
  n_initial,
  scale_inputs,
  inf_crit
)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{lags}{A list containing integer vectors with the lags associated with each output variable.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data \code{period = c(12)}, for hourly data \code{period = c(24, 168)}).}

\item{const}{Logical value. If \code{TRUE}, a constant term (intercept) is used.}

\item{n_sdiff}{Integer vector. The number of seasonal differences.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("AIC", "BIC", "HQ")}.}
}
\value{
A list with vectors const and lags
}
\description{
Model selection procedure
}
