% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lags}
\alias{select_lags}
\title{Estimate best subset autoregressive models and select lags}
\usage{
select_lags(.data, period, n_sdiff, n_diff, max_lag)
}
\arguments{
\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer vector. The number of seasonal differences.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}

\item{max_lag}{Integer value. Maximum number of non-seasonal lags.}

\item{data}{Numeric matrix containing the original data.}
}
\value{

}
\description{
Estimate best subset autoregressive models and select lags.
   The underlying model is an ARp model (subset autoregressive model) and
   by default the BICq information criterion is used. This is a wrapper
   function for \code{select_lags_vec} that works on the columns of a
   numeric matrix.
}
