% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lags_vec}
\alias{select_lags_vec}
\title{Estimate and select significant lags by partial autocorrelation function (PACF).}
\usage{
select_lags_vec(y, period, n_sdiff, n_diff, level = 0.9)
}
\arguments{
\item{y}{Numeric vector containing the original data.}

\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer value. The number of seasonal differences.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}

\item{level}{Numeric value. The confidence level.}
}
\value{

}
\description{
Estimate and select significant lags by partial autocorrelation function (PACF).
}
