% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulate_esn}
\alias{simulate_esn}
\title{Simulate a fitted ESN.}
\usage{
simulate_esn(
  win,
  wres,
  wout,
  n_ahead,
  alpha,
  lags,
  inputs,
  states_train,
  error,
  n_sim
)
}
\arguments{
\item{win}{Numeric matrix. Weights for the input variables.}

\item{wres}{Numeric matrix. Weights for the reservoir.}

\item{wout}{Numeric matrix. Weights for the output variables (estimated coefficients from ridge regression).}

\item{n_ahead}{Integer value. The number of periods for forecasting (forecast horizon).}

\item{alpha}{Numeric value. The Leakage rate (smoothing parameter).}

\item{lags}{List containing integer vectors with the lags associated with each output variable.}

\item{inputs}{Numeric matrix. Initialized input features (gets updated during forecasting process).}

\item{states_train}{Numeric matrix. Internal states from training (necessary due to last values).}

\item{error}{Numeric matrix. The innovations for simulation (re-sampled residuals from fitted model).}

\item{n_sim}{Integer value. The number of simulations.}
}
\value{
sim List with simulated future sample paths as numeric matrix.
}
\description{
Simulate future sample path from a fitted ESN.
}
