% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{summary.esn}
\alias{summary.esn}
\title{Provide a detailed summary of the trained ESN model}
\usage{
\method{summary}{esn}(object, ...)
}
\arguments{
\item{object}{An object of class \code{esn}. The result of a call to \code{train_esn()}.}

\item{...}{Currently not in use.}
}
\value{
Print detailed model summary.
}
\description{
Provide a detailed summary of the trained ESN model.
}
\examples{
xdata <- as.numeric(AirPassengers)
xmodel <- train_esn(y = xdata)
summary(xmodel)

}
\seealso{
Other base functions: 
\code{\link{forecast_esn}()},
\code{\link{is.esn}()},
\code{\link{is.forecast_esn}()},
\code{\link{is.tune_esn}()},
\code{\link{plot.esn}()},
\code{\link{plot.forecast_esn}()},
\code{\link{plot.tune_esn}()},
\code{\link{print.esn}()},
\code{\link{summary.tune_esn}()},
\code{\link{train_esn}()},
\code{\link{tune_esn}()}
}
\concept{base functions}
