% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_esn.R
\name{theme_esn}
\alias{theme_esn}
\title{theme_esn}
\usage{
theme_esn(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{Base font size}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}

\item{base_familiy}{Base font family}
}
\description{
Custom ggplot2 theme for package "esn"
}
