% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARX.R
\name{tidy.ARX}
\alias{tidy.ARX}
\title{Estimated coefficients}
\usage{
\method{tidy}{ARX}(object)
}
\arguments{
\item{object}{An object of class \code{ARX}.}
}
\value{
Coefficients extracted from the object.
}
\description{
Return the estimated coefficients from a trained ARX as tibble.
}
