% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{train}
\alias{train}
\title{Train ESN.}
\usage{
train(.data, specials, ...)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_esn}.}

\item{...}{Further arguments passed to \code{echos::auto_esn(...)}.}
}
\value{
An object of class \code{fbl_esn}.
}
\description{
Train an Echo State Network (ESN).
}
