% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_arx.R
\name{train_arx}
\alias{train_arx}
\title{Train an Autoregressive model}
\usage{
train_arx(
  data,
  lags,
  fourier = NULL,
  xreg = NULL,
  dy = 0,
  dx = 0,
  inf_crit = "aic",
  n_initial = 10,
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{data}{A \code{tsibble} containing the response variable.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{inf_crit}{Character value. The information criterion used for variable selection \code{inf_crit = c("aic", "aicc", "bic")}.}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
A \code{list} containing:
   \itemize{
      \item{\code{actual}: Numeric vector containing the actual values.}
      \item{\code{fitted}: Numeric vector containing the fitted values.}
      \item{\code{resid}: Numeric vector containing the residuals.}
      \item{\code{method}: A \code{list} containing several objects and meta information of the trained model (coefficients, model metrics, etc.).}
      }
}
\description{
This function trains an Autoregressive model (ARX) to a
  univariate time series.
}
