% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_glmnet.R
\name{train_glmnet}
\alias{train_glmnet}
\title{Estimate a GLM via elastic net regularization}
\usage{
train_glmnet(X, y, lambda, type, weights, penalty, ...)
}
\arguments{
\item{X}{Numeric matrix. The design matrix containing the predictor variables.}

\item{y}{Numeric matrix. The response variable(s).}

\item{lambda}{Numeric value. The regularization parameter.}

\item{type}{Numeric value. The elastic net mixing parameter.}

\item{weights}{Numeric vector. Observation weights.}

\item{...}{Further arguments passed to glmnet::glmnet().}
}
\value{
A list containing the estimated coefficients, fitted values etc.
}
\description{
Estimate a Generalized Linear Model (GLM) via elastic net
  regularization.
}
