% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{train_ridge}
\alias{train_ridge}
\title{Estimate ridge regression}
\arguments{
\item{X}{Numeric matrix. The design matrix containing the predictor variables.}

\item{y}{Numeric matrix. The response variable(s).}

\item{lambda}{Numeric value. The regularization parameter.}

\item{weights}{Numeric vector. Observation weights.}
}
\value{
A list containing the estimated coefficients, fitted values, etc.
}
\description{
Estimate coefficients via ridge regression.
}
