% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_inputs.R
\name{tune_inputs}
\alias{tune_inputs}
\title{Random subset regression to select model inputs}
\usage{
tune_inputs(
  data,
  lags,
  fourier,
  xreg,
  dy,
  dx,
  n_initial,
  scale_inputs,
  inf_crit,
  n_models,
  n_seed
)
}
\arguments{
\item{data}{A \code{tsibble} containing the response variable.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{inf_crit}{Character value. The information criterion used for variable selection \code{inf_crit = c("aic", "aicc", "bic")}.}

\item{n_models}{Integer value. The maximum number of (random) models to train for model selection.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}
}
\value{
A \code{list} containing:
  \itemize{
    \item{\code{lags}: A \code{list} containing integer vectors with the lags associated with each input variable.}
    \item{\code{fourier}: Integer vector. The fourier terms (seasonal cycles per period).}
    }
}
\description{
The function \code{tune_inputs()} performs random subset
  regression to identify the optimal model inputs. The models are estimated
  via Ordinary Least Squares (OLS).
}
