% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_pars.R
\name{tune_pars}
\alias{tune_pars}
\title{Estimate information criterion for hyperparameter tuning.}
\usage{
tune_pars(
  data,
  pars,
  lags,
  n_fourier,
  period,
  const = TRUE,
  n_diff = 0,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  density = 0.1,
  inf_crit = "bic",
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{pars}{Numeric vector containing the hyperparameters.}

\item{lags}{A list containing integer vectors with the lags associated with each output variable.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data \code{period = c(12)}, for hourly data \code{period = c(24, 168)}).}

\item{const}{Logical value. If \code{TRUE}, a constant term (intercept) is used.}

\item{n_diff}{Integer value. The number of differences.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("aic", "bic", "hq")}.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
out Numeric value. The information criterion to be minimized.
}
\description{
This function estimates the information criterion for hyperparameter
   tuning. The function returns a numeric value (= information criterion), which
   is then minimized within a call to \code{optim()} for varying hyperparameters.
}
