% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{auto_esn}
\alias{auto_esn}
\title{Automatic train an Echo State Network}
\usage{
auto_esn(.data, specials, ...)
}
\arguments{
\item{.data}{A \code{tsibble} containing the time series data.}

\item{specials}{Currently not is use.}

\item{...}{Further arguments passed to \code{train_esn()}.}
}
\value{
An object of class \code{ESN}.
}
\description{
This function trains an Echo State Network (ESN)
  to a univariate time series.
}
