% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{auto_esn}
\alias{auto_esn}
\title{Automatic train an Echo State Network (ESN)}
\usage{
auto_esn(
  .data,
  specials,
  const = TRUE,
  lags = NULL,
  n_fourier = NULL,
  n_initial = 10,
  n_res = 200,
  n_seed = 42,
  alpha = 0.8,
  rho = 1,
  lambda = 1,
  density = 0.1,
  scale_inputs = c(-1, 1),
  scale_runif = c(-0.5, 0.5),
  inf_crit = "bic",
  control_inputs = list(tune = TRUE, n_sample = 1000),
  control_pars = list(tune = TRUE, lower = c(0, 0.1, 0.1), upper = c(1, 5, 10)),
  ...
)
}
\arguments{
\item{.data}{A \code{tsibble} containing the time series data.}

\item{specials}{Currently not is use.}

\item{const}{Logical value. If \code{TRUE}, a constant term (intercept) is used.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}

\item{rho}{Numeric value. The spectral radius for scaling the reservoir weight matrix.}

\item{lambda}{Numeric value. The regularization (shrinkage) parameter for ridge regression.}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("aic", "bic", "hq")}.}

\item{control_inputs}{A \code{list} containing control values for the automatic selection of model inputs:
\itemize{
  \item{\code{tune}: Logical value. If \code{TRUE}, the model inputs are tuned, otherwise model inputs are used as defined.}
  \item{\code{n_sample}: Integer value. The number of samples for the random grid.}
     }}

\item{control_pars}{A \code{list} containing control values for the hyperparameter tuning:
\itemize{
  \item{\code{tune}: Logical value. If \code{TRUE}, the hyperparameters are tuned, otherwise hyperparameters are used as defined.}
  \item{\code{lower}: Numeric vector. The lower bounds for \code{alpha}, \code{rho} and \code{lambda} used in the optimization.}
  \item{\code{upper}: Numeric vector. The upper bounds for \code{alpha}, \code{rho} and \code{lambda} used in the optimization.}
     }}

\item{...}{Further arguments passed to \code{stats::optim()}}
}
\value{

}
\description{
This function trains an Echo State Network (ESN)
  to a univariate time series.
}
