% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_esn.R
\name{auto_esn}
\alias{auto_esn}
\title{Automatic train an Echo State Network (ESN)}
\usage{
auto_esn(
  data,
  period,
  max_lag = 6,
  n_fourier = NULL,
  n_initial = 10,
  n_res = 200,
  density = 0.1,
  scale_inputs = c(-1, 1),
  inf_crit = "BIC",
  n_sample = 5000
)
}
\arguments{
\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data \code{period = c(12)}, for hourly data \code{period = c(24, 168)}).}

\item{max_lag}{Integer value. Maximum number of non-seasonal lags.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("AIC", "BIC", "HQ")}.}

\item{n_sample}{Integer value. The number of random samples for random search.}

\item{.data}{A \code{tsibble} containing the time series data.}
}
\value{
An object of class \code{ESN}.
}
\description{
This function trains automatically an Echo State Network (ESN)
   to a univariate or multivariate time series.
}
