% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{auto_esn}
\alias{auto_esn}
\title{Automatic train an Echo State Network}
\usage{
auto_esn(
  .data,
  specials,
  lags = NULL,
  fourier = NULL,
  xreg = NULL,
  dy = 0,
  dx = 0,
  inf_crit = "aic",
  n_states = 500,
  n_models = 500,
  n_seed = 42,
  alpha = 1,
  rho = 1,
  density = 0.1,
  scale_win = 0.5,
  scale_wres = 0.5,
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{.data}{A \code{tsibble} containing the time series data.}

\item{specials}{Currently not is use.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{inf_crit}{Character value. The information criterion used for variable selection \code{inf_crit = c("aic", "aicc", "bic")}.}

\item{n_states}{Integer value. The number of internal states per reservoir.}

\item{n_models}{Integer value. The maximum number of (random) models to train for model selection.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}

\item{rho}{Numeric value. The spectral radius for scaling the reservoir weight matrix.}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_win}{Numeric value. The lower and upper bound of the uniform distribution for scaling the input weight matrix.}

\item{scale_wres}{Numeric value. The lower and upper bound of the uniform distribution for scaling the reservoir weight matrix.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
An object of class \code{ESN}.
}
\description{
This function trains an Echo State Network (ESN)
  to a univariate time series.
}
