% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_esn.R
\name{auto_esn}
\alias{auto_esn}
\title{auto_esn}
\usage{
auto_esn(
  data,
  n_res = 200,
  n_initial = 10,
  n_samples = 200,
  n_seed = 42,
  max_lag = 18,
  max_cycles = c(12, 12),
  diff = FALSE,
  scale_inputs = c(-2, 2),
  scale_runif = c(-2, 2),
  trans_inputs = FALSE
)
}
\arguments{
\item{data}{A tsibble containing the time series data. Must have column "time" with time index (date or date-time).}

\item{n_samples}{Integer value. The number of samples for the random grid.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{max_lag}{Integer value. The maximum number of non-seasonal lags.}

\item{diff}{Logical value. If \code{TRUE}, the time series is modeled in first difference.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{trans_inputs}{Logical value. If TRUE, the data are transformed using ordered quantile normalizing transformation (OQR).}

\item{max_cycle}{Integer vector. The maximum number of trigonometric terms per season.}
}
\value{
An object of class \code{esn}.
}
\description{
auto_esn ...
}
