% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_revolved}
\alias{create_revolved}
\title{Create lagged variables of a matrix for iterative forecasting}
\usage{
create_revolved(data, lags, n_ahead)
}
\arguments{
\item{data}{Numeric vector or matrix. Each column is a variable and each row an observation.}

\item{lags}{List containing vectors with the number of lags (in units of observations) per variable.}

\item{n_ahead}{Integer value. The forecast horizon (n-step ahead).}
}
\value{
y_lag Numeric matrix with the lagged variables of the input data for iterative forecasting.
}
\description{
Create lagged variables of a matrix for iterative forecasting,
  shifting each column back by a given number of observations and fill with NAs for the updates.
}
