% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_spec}
\alias{create_spec}
\title{Create model specification.}
\usage{
create_spec(n_layers, pars, n_trig, period)
}
\arguments{
\item{n_layers}{List containing the number of inputs (n_inputs), reservoir size (n_res) and the number of outputs (n_outputs).}

\item{pars}{List containing the hyperparameters alpha, rho, lambda and density.}

\item{n_trig}{Integer vector. The number of seasonal cycles per period.}

\item{period}{Integer vector. The periodicity of the time series.}
}
\value{
model_spec Character value. The model specification as string.
}
\description{
This function creates the model specification (short summary) as a string.
}
