% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_win}
\alias{create_win}
\title{Create the input weight matrix}
\usage{
create_win(n_inputs, n_res, scale_runif)
}
\arguments{
\item{n_inputs}{Integer value. The number of input features.}

\item{n_res}{Integer value. The number of internal states within the reservoir (reservoir size).}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}
}
\value{
win List containing the input weight matrices.
}
\description{
This function creates the random input weight matrices.
}
