% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_wres}
\alias{create_wres}
\title{Create the reservoir weight matrix}
\usage{
create_wres(n_res, rho, density, scale_runif, symmetric = FALSE)
}
\arguments{
\item{n_res}{Integer value. The number of internal states within the reservoir (reservoir size).}

\item{rho}{Numeric value. The spectral radius for scaling the weight matrix.}

\item{density}{Numeric value. The parameter defines the connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{symmetric}{Logical value. If \code{TRUE}, the matrix is symmetric.}
}
\value{
wres Numeric matrix. The final reservoir weight matrix.
}
\description{
This function creates the random reservoir weight matrix
   (scaled to spectral radius rho).
}
