% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_data.R
\name{diff_data}
\alias{diff_data}
\title{Calculate n-th differences}
\usage{
diff_data(data, n_diff, na_rm = TRUE)
}
\arguments{
\item{data}{}

\item{na_rm}{}
}
\value{

}
\description{
This function takes a numeric matrix and calculates n-th differences for each column. Leading NAs are padded.
}
