% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_data}
\alias{diff_data}
\title{Calculate seasonal and non-seasonal differences of a numeric matrix.}
\usage{
diff_data(data, period, n_sdiff, n_diff)
}
\arguments{
\item{data}{Numeric matrix.}

\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer vector. The number of seasonal differences.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}
}
\value{
y_diff Numeric matrix with the differenced data.
}
\description{
This function takes a numeric matrix and calculates seasonal and non-seasonal differences for each column.
}
