% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_vec}
\alias{diff_vec}
\title{Calculate (non-seasonal) differences of a numeric vector}
\usage{
diff_vec(y, n_diff)
}
\arguments{
\item{y}{Numeric vector.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}
}
\value{
y_diff Numeric vector with the differenced data.
}
\description{
This function takes a numeric vector and calculates
   (non-seasonal) differences.
}
