% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{ESN}
\alias{ESN}
\title{Automatic training of ESNs}
\usage{
ESN(formula, ...)
}
\arguments{
\item{formula}{Model specification (see "Specials" section, currently not in use...)}

\item{...}{Further arguments passed to \code{echos::auto_esn(...)}.}
}
\value{
esn_model An object of class \code{fbl_esn}.
}
\description{
Automatic training of ESNs
}
