% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{ESN}
\alias{ESN}
\title{Automatic train an Echo State Network}
\usage{
ESN(formula, ...)
}
\arguments{
\item{formula}{Model specification (currently not in use).}

\item{...}{Further arguments passed to \code{train_esn()}.}
}
\value{
An object of class \code{ESN}.
}
\description{
This function trains an Echo State Network (ESN)
  to a univariate time series.
}
\examples{
library(tsibble)
library(fable)
AirPassengers \%>\%
as_tsibble() \%>\%
model("ESN" = ESN(value))

}
