% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{extract_esn}
\alias{extract_esn}
\title{Extract and return values from a trained ESN as tibble}
\usage{
extract_esn(object, ...)
}
\arguments{
\item{object}{An object of class \code{mdl_df}.}
}
\value{
A tibble containing the hyper-parameters.
}
\description{
Extract and return values from a trained ESN
  as tibble. The function works only if the
  model within the \code{mdl_df} is of class \code{ESN}.
  The extracted values are are stored in a tibble with
  the following columns:
  
  \itemize{
    \item{\code{spec}: Character value. Succinct summary of model specifications.}
    \item{\code{n_inputs}: Integer value. The number of model inputs.}
    \item{\code{n_res}: Integer value. The number of internal states within the reservoir (hidden layer).}
    \item{\code{n_outputs}: Integer value. The number of model outputs.}
    \item{\code{n_diff}: Integer value. The number of non-seasonal differences.}
    \item{\code{const}: Logical value. If \code{TRUE}, a constant term (intercept) is used.}
    \item{\code{lags}: A \code{list} containing integer vectors with the lags associated with each input variable.}
    \item{\code{period}: A \code{list} containing the periodicity as integer vector.}
    \item{\code{alpha}: Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}
    \item{\code{rho}: Numeric value. The spectral radius for scaling the reservoir weight matrix.}
    \item{\code{lambda}: Numeric value. The regularization (shrinkage) parameter for ridge regression.}
    \item{\code{density}: Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}
    \item{\code{df}: Numeric value. The effective degree of freedom.}
    \item{\code{aic}: Numeric value. The Akaike information criterion.}
    \item{\code{bic}: Numeric value. The Bayesian information criterion.}
    \item{\code{hq}: Numeric value. The Hannan-Quinn information criterion.}
    }
}
