% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{fitted.ESN}
\alias{fitted.ESN}
\title{Extract fitted values from a trained ESN.}
\usage{
fitted.ESN(object, ...)
}
\arguments{
\item{object}{The time series model used to produce the forecast.}

\item{...}{Additional arguments passed to ...}
}
\value{

}
\description{
Extract fitted values from a trained ESN.
}
