% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{fitted.ESN}
\alias{fitted.ESN}
\title{Extract fitted values from a trained ESN}
\usage{
\method{fitted}{ESN}(object)
}
\arguments{
\item{object}{An object of class \code{ESN}.}
}
\value{

}
\description{
Extract fitted values from a trained ESN.
}
