% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{forecast.ESN}
\alias{forecast.ESN}
\title{Forecast a trained ESN}
\usage{
\method{forecast}{ESN}(object, new_data, specials = NULL, xreg = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ESN}.}

\item{new_data}{Forecast horizon (n-step ahead forecast)}

\item{specials}{Currently not in use}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{...}{Currently not in use.}
}
\value{
An object of class \code{fable}.
}
\description{
Forecast a trained ESN.
}
\examples{
library(tsibble)
library(fable)
AirPassengers \%>\%
as_tsibble() \%>\%
model("ESN" = ESN(value)) \%>\%
forecast(h = 18)

}
