% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_esn.R
\name{forecast_esn}
\alias{forecast_esn}
\title{Forecast an Echo State Network}
\usage{
forecast_esn(object, n_ahead = 12, n_sim = 100, n_seed = 42)
}
\arguments{
\item{object}{An object of class "esn". The result of a call to train_esn(...).}

\item{n_ahead}{Integer value. The number of periods for forecasting (forecast horizon).}

\item{n_sim}{Integer value. The number of simulations (number of future sample paths).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}
}
\value{
forecast A tibble containing the forecast and prediction intervals with the columns time, id, type, variable, and value.

simulation A tibble containing the simulations with the columns time, id, type (sim(1), sim(2), ...), variable, and value.

actual A tibble containing the actual values with the columns time, id, type, variable, and value.

fitted A tibble containing the fitted values with the columns time, id, type, variable, and value.

states_fcst A tibble containing the internal states with the columns time, id, type, variable, and value.

method A list containing several objects and information of the trained ESN (weight matrices, hyperparameters, etc.).
}
\description{
Forecast an Echo State Network from a trained model (univariate and multivariate ESNs)
}
