% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_esn.R
\name{forecast_esn}
\alias{forecast_esn}
\title{Forecast a trained Echo State Network (ESN).}
\usage{
forecast_esn(object, n_ahead = 12, n_sim = 100, n_seed = 42)
}
\arguments{
\item{object}{An object of class \code{ESN}. The result of a call to \code{train_esn()} or \code{auto_esn()}.}

\item{n_ahead}{Integer value. The number of periods for forecasting (forecast horizon).}

\item{n_sim}{Integer value. The number of simulations (number of future sample paths).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}
}
\value{
A \code{list} containing:
   \itemize{
      \item{\code{forecast}: A \code{tsibble} containing the forecasts.}
      \item{\code{simulation}: A \code{tsibble} containing the simulated future sample path.}
      \item{\code{states_train}: A \code{tsibble} containing the internal states.}
      \item{\code{method}: A list containing several objects and information of the trained ESN (weight matrices, hyperparameters, model metrics, etc.).}
      \item{\code{n_ahead}: Integer value. The number of periods for forecasting (forecast horizon).}
      \item{\code{n_sim}: Integer value. The number of simulations (number of future sample paths).}
      }
}
\description{
Forecast an Echo State Network from a trained model.
}
