% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{glance.ESN}
\alias{glance.ESN}
\title{Summary of model fit}
\usage{
\method{glance}{ESN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ESN}.}

\item{...}{Currently not in use.}
}
\value{
Summary statistics extracted from the object.
}
\description{
Return summary statistics from a trained ESN as tibble.
 \itemize{
   \item{\code{df}: Effective degrees of freedom.}
   \item{\code{aic}: Akaike information criterion.}
   \item{\code{bic}: Bayesian information criterion.}
   \item{\code{hqc}: Hannan-Quinn information criterion.}
      }
}
