% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_diff_data}
\alias{inv_diff_data}
\title{Integrate (non-seasonal) differences of a numeric matrix
   ("inverse difference")}
\usage{
inv_diff_data(data, data_diff, n_diff)
}
\arguments{
\item{data}{Numeric matrix containing the original data.}

\item{data_diff}{Numeric matrix containing the differenced data.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}
}
\value{
y_int Numeric matrix with the inverse differenced data.
}
\description{
This function takes a numeric matrix and integrates
   (non-seasonal) differences for each column ("inverse difference").
}
