% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_trans_data}
\alias{inv_trans_data}
\title{Inverse transforms the columns of a numeric matrix.}
\usage{
inv_trans_data(object, new_data)
}
\arguments{
\item{object}{List with objects of class \code{orderNorm}.}

\item{new_data}{Numeric vector. New data for inverse transforming.}
}
\value{
data Numeric matrix. Inverse transformed data.
}
\description{
Inverse transforms the columns of a numeric matrix using inverse ordered quantile normalizing transformation (OQR).
}
