% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{is.forecast_esn}
\alias{is.forecast_esn}
\title{Checks if object is of class "forecast_esn"}
\usage{
is.forecast_esn(object)
}
\arguments{
\item{object}{object to be tested.}
}
\value{
Logical value. If \code{TRUE}, the object is of class \code{forecast_esn}.
}
\description{
Returns \code{TRUE} if the object is of class \code{forecast_esn}.
}
\examples{
xdata <- as.numeric(AirPassengers)
xmodel <- train_esn(y = xdata)
xfcst <- forecast_esn(xmodel, n_ahead = 12)
is.forecast_esn(xfcst)

}
\seealso{
Other base functions: 
\code{\link{forecast_esn}()},
\code{\link{is.esn}()},
\code{\link{is.tune_esn}()},
\code{\link{plot.esn}()},
\code{\link{plot.forecast_esn}()},
\code{\link{plot.tune_esn}()},
\code{\link{print.esn}()},
\code{\link{summary.esn}()},
\code{\link{summary.tune_esn}()},
\code{\link{train_esn}()},
\code{\link{tune_esn}()}
}
\concept{base functions}
