% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{is.forecast_esn}
\alias{is.forecast_esn}
\title{Checks if object is of class "forecast_esn"}
\usage{
is.forecast_esn(object)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Returns \code{TRUE} if the object is of class "forecast_esn".
}
