% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4_data}
\alias{m4_data}
\title{M4 dataset}
\format{
A time series object of class \code{tsibble} with 1.152 rows and 4 columns:
   \itemize{
      \item{\code{series}: Unique identifier as \code{character} (key variable).}
      \item{\code{category}: Category (e.g., Demographic, Macro) as \code{factor}.}
      \item{\code{index}: Date as \code{yearmonth} (index variable).}
      \item{\code{value}: Value as \code{numeric} (measurement variable).}
      }
}
\source{
\href{https://github.com/Mcompetitions/M4-methods}{M4 Forecasting Competition}
}
\usage{
data(m4_data)
}
\description{
Monthly tsibble with six exemplary time series from the M4 
  Forecasting Competition.
}
\examples{
data(m4_data)
}
\keyword{datasets}
