% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-fable.R
\name{model_sum.ESN}
\alias{model_sum.ESN}
\title{Model specification of a trained ESN model}
\usage{
\method{model_sum}{ESN}(x)
}
\arguments{
\item{x}{An object of class \code{mdl_df}, containing an ESN model.}
}
\value{
Model summary extracted from the object.
}
\description{
Provides a compact overview of the model specification in the 
   format \code{ESN({n_states, alpha, rho}, {n_models, df})}.
}
\examples{
library(tsibble)
library(fable)
AirPassengers \%>\%
as_tsibble() \%>\%
model("ESN" = ESN(value))

}
