% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{model_sum.ESN}
\alias{model_sum.ESN}
\title{Provide a succinct summary of a trained ESN.}
\usage{
\method{model_sum}{ESN}(x)
}
\arguments{
\item{object}{The ESN to summarize.}
}
\value{

}
\description{
Provide a succinct summary of a trained ESN.
}
