% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{plot.esn}
\alias{plot.esn}
\title{Plot internal states of a trained ESN model.}
\usage{
\method{plot}{esn}(x, ...)
}
\arguments{
\item{x}{An object of class \code{esn}.}

\item{...}{Currently not in use.}
}
\value{
Line chart of internal states.
}
\description{
Plot internal states (i.e., the reservoir) of a trained ESN 
  model as line chart.
}
\examples{
xdata <- as.numeric(AirPassengers)
xmodel <- train_esn(y = xdata)
plot(xmodel)

}
