% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_esn.R
\name{predict_esn}
\alias{predict_esn}
\title{Forecast a fitted ESN}
\usage{
predict_esn(
  win,
  wres,
  wout,
  n_ahead,
  alpha,
  lags,
  inputs,
  states_train,
  innov = NULL
)
}
\arguments{
\item{win}{Weights matrix for input variables}

\item{wres}{Weights matrix for reservoir}

\item{wout}{Weights matrix for output variables (coefficients from ridge regression)}

\item{n_ahead}{Number of periods for forecasting (forecast horizon)}

\item{alpha}{Leakage rate (smoothing parameter)}

\item{lags}{A numeric matrix with lagged variables}

\item{inputs}{Initialized input features (gets updated during forecasting process)}

\item{states_train}{Internal states from training (necessary for last values)}

\item{innov}{Innovations for a simulation (default is NULL)}
}
\value{
fcst A numeric matrix with point forecasts

states_fcst A numeric matrix with internal states used for forecasting
}
\description{
Calculate point forecasts for a fitted ESN
}
