% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{residuals.ESN}
\alias{residuals.ESN}
\title{Extract residuals from a trained ESN}
\usage{
\method{residuals}{ESN}(object)
}
\arguments{
\item{object}{An object of class \code{ESN}.}
}
\value{

}
\description{
Extract residuals from a trained ESN.
}
