% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lags}
\alias{select_lags}
\title{Estimate and select significant lags by partial autocorrelation function (PACF).}
\usage{
select_lags(.data, period, n_sdiff, n_diff, level = 0.9)
}
\arguments{
\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer vector. The number of seasonal differences.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}

\item{level}{Numeric value. The confidence level.}

\item{data}{Numeric matrix containing the original data.}
}
\value{

}
\description{
Estimate and select significant lags by partial autocorrelation function (PACF).
}
