% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_lags_vec}
\alias{select_lags_vec}
\title{Estimate best subset autoregressive models and select lags}
\usage{
select_lags_vec(y, period, n_sdiff, n_diff, max_lag)
}
\arguments{
\item{y}{Numeric vector containing the original data.}

\item{period}{Integer vector. The periodicity of the time series.}

\item{n_sdiff}{Integer value. The number of seasonal differences.}

\item{n_diff}{Integer value. The number of non-seasonal differences.}

\item{max_lag}{Integer value. Maximum number of non-seasonal lags.}
}
\value{

}
\description{
Estimate best subset autoregressive models and select lags.
   The underlying model is an ARp model (subset autoregressive model) and
   by default the BICq information criterion is used.
}
