% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_esn.R
\name{simulate_esn}
\alias{simulate_esn}
\title{Simulate a fitted ESN}
\usage{
simulate_esn(
  win,
  wres,
  wout,
  n_ahead,
  alpha,
  lags,
  inputs,
  states_train,
  error,
  n_sim
)
}
\arguments{
\item{win}{Weights matrix for input variables}

\item{wres}{Weights matrix for reservoir}

\item{wout}{Weights matrix for output variables (coefficients from ridge regression)}

\item{n_ahead}{Number of periods for forecasting (forecast horizon)}

\item{alpha}{Leakage rate (smoothing parameter)}

\item{lags}{A numeric matrix with lagged variables}

\item{inputs}{Initialized input features (gets updated during forecasting process)}

\item{states_train}{Internal states from training (necessary for last values)}

\item{error}{Residuals from fitted model}

\item{n_sim}{Number of simulations}
}
\value{
sim An array with simulated future sample paths
}
\description{
Simulate future sample path from a fitted ESN.
}
