% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{summary.esn}
\alias{summary.esn}
\title{Provide a detailed summary of the trained ESN model}
\usage{
\method{summary}{esn}(object, ...)
}
\arguments{
\item{object}{An object of class \code{esn}.}

\item{...}{Currently not in use.}
}
\value{
Print detailed model summary.
}
\description{
Provide a detailed summary of the trained ESN model.
}
\examples{
xdata <- as.numeric(AirPassengers)
xmodel <- train_esn(y = xdata)
summary(xmodel)
}
