% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_data}
\alias{synthetic_data}
\title{Synthetic data}
\format{
An object of class \code{tibble} with 2.000 rows and 3 columns:
   \itemize{
      \item{\code{variable}: Unique identifier as \code{character} (key variable).}
      \item{\code{index}: Index as \code{inteher} (index variable).}
      \item{\code{value}: Value as \code{numeric} (measurement variable).}
      }
}
\usage{
data(synthetic_data)
}
\description{
\code{tibble} with ten synthetic time series, i.e., Square Wave,
  Sawtooth Wave, Harmonic Wave, Harmonic Wave w/ Trend, Amplitude Modulated
  Wave, Frequency Modulated Wave, AR(1) Process, MA(2) Process, White Noise
  Process and Random Walk Process
}
\examples{
data(synthetic_data)
}
\keyword{datasets}
