% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{tidy.ESN}
\alias{tidy.ESN}
\title{Estimated coefficients}
\usage{
\method{tidy}{ESN}(object)
}
\arguments{
\item{object}{An object of class \code{ESN}.}
}
\value{
A tibble containing the estimated coefficients.
}
\description{
Return the estimated coefficients from a trained ESN as tibble.
}
