% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{tidy.ESN}
\alias{tidy.ESN}
\title{Estimated coefficients}
\usage{
\method{tidy}{ESN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ESN}.}

\item{...}{Currently not in use.}
}
\value{
Coefficients extracted from the object.
}
\description{
Return the estimated coefficients from a trained ESN as tibble.
}
\examples{
library(tsibble)
library(fable)
AirPassengers \%>\%
as_tsibble() \%>\%
model("ESN" = ESN(value)) \%>\%
tidy()

}
