% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESN.R
\name{train_esn}
\alias{train_esn}
\title{Train an Echo State Network (ESN).}
\usage{
train_esn(
  .data,
  specials,
  lags,
  n_trig,
  period,
  const = TRUE,
  n_diff = 0,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  density = 0.1,
  scale_inputs = c(-1, 1),
  inf_crit = "HQ",
  ...
)
}
\arguments{
\item{.data}{Input data as tsibble.}

\item{specials}{Specials as list defined in \code{specials_esn}.}

\item{...}{Further arguments passed to \code{echos::auto_esn(...)}.}
}
\value{
An object of class \code{ESN}.
}
\description{
Train an Echo State Network (ESN).
}
