% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_esn.R
\name{train_esn}
\alias{train_esn}
\title{Train an Echo State Network (ESN).}
\usage{
train_esn(
  data,
  lags,
  n_fourier,
  period,
  const = TRUE,
  n_sdiff = 0,
  n_diff = 0,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  rho = 0.7,
  alpha = 0.85,
  lambda = 12,
  density = 0.1,
  scale_runif = c(-0.5, 0.5),
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{lags}{A list containing integer vectors with the lags associated with each output variable.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data \code{period = c(12)}, for hourly data \code{period = c(24, 168)}).}

\item{const}{Logical value. If \code{TRUE}, a constant term (intercept) is used.}

\item{n_sdiff}{Integer vector. The number of seasonal differences.}

\item{n_diff}{Integer vector. The number of non-seasonal differences.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{rho}{Numeric value. The spectral radius for scaling the reservoir weight matrix.}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}

\item{lambda}{Numeric value. The regularization (shrinkage) parameter for ridge regression.}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
data The original input data as \code{tsibble}.

actual A \code{tsibble} containing the actual values in long format.

fitted A \code{tsibble} containing the fitted values in long format.

error A \code{tsibble} containing the errors (= residuals, i.e. actual - fitted) in long format.

states_train A \code{tsibble} containing the internal states in long format.

method A list containing several objects and information of the trained ESN (weight matrices, hyperparameters, model metrics, etc.).
}
\description{
This function trains an Echo State Network (ESN) to a univariate or multivariate time series.
}
