% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_esn.R
\name{train_esn}
\alias{train_esn}
\title{Train an Echo State Network}
\usage{
train_esn(
  data,
  lags,
  fourier = NULL,
  xreg = NULL,
  dy = 0,
  dx = 0,
  n_models = 500,
  inf_crit = "aic",
  max_states = 30,
  n_best = 50,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  alpha = 0.8,
  rho = 1,
  density = 0.1,
  scale_win = 0.1,
  scale_wres = 0.5,
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{data}{A \code{tsibble} containing the response variable.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}

\item{rho}{Numeric value. The spectral radius for scaling the reservoir weight matrix.}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_win}{Numeric value. The lower and upper bound of the uniform distribution for scaling the input weight matrix.}

\item{scale_wres}{Numeric value. The lower and upper bound of the uniform distribution for scaling the reservoir weight matrix.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
A \code{list} containing:
   \itemize{
      \item{\code{actual}: Numeric vector containing the actual values.}
      \item{\code{fitted}: Numeric vector containing the fitted values.}
      \item{\code{resid}: Numeric vector containing the residuals.}
      \item{\code{states_train}: Numeric matrix containing the internal states.}
      \item{\code{method}: A \code{list} containing several objects and meta information of the trained ESN (weight matrices, hyperparameters, model metrics, etc.).}
      }
}
\description{
This function trains an Echo State Network (ESN) to a
  univariate time series.
}
