% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_esn.R
\name{train_esn}
\alias{train_esn}
\title{Train an Echo State Network (ESN).}
\usage{
train_esn(
  data,
  lags,
  n_trig,
  period,
  const = TRUE,
  n_diff = 0,
  n_res = 200,
  n_initial = 10,
  n_seed = 42,
  rho = 0.7,
  alpha = 0.85,
  lambda = 12,
  density = 0.1,
  scale_runif = c(-0.5, 0.5),
  scale_inputs = c(-1, 1)
)
}
\arguments{
\item{data}{A tsibble containing the time series data. Must have column "time" with time index (date or date-time).}

\item{lags}{A list containing integer vectors with the lags associated with each output variable.}

\item{n_trig}{Integer vector. The number of seasonal cycles per period.}

\item{period}{Integer vector. The periodicity of the time series (e.g. for monthly data period = c(12), for hourly data period = c(24, 168)).}

\item{const}{Logical value. If TRUE, a constant term (intercept) is used.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{rho}{Numeric value. The spectral radius for scaling the reservoir weight matrix.}

\item{alpha}{Numeric value. The leakage rate (smoothing parameter) applied to the reservoir.}

\item{lambda}{Numeric value. The regularization (shrinkage) parameter for ridge regression.}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{diff}{Logical value. If TRUE, the time series is modeled in first difference.}
}
\value{
data The original input data as tibble.

actual A tibble containing the actual values with the columns time, id, type, variable, and value.

fitted A tibble containing the fitted values with the columns time, id, type, variable, and value.

error A tibble containing the errors (= residuals, i.e. actual - fitted) with the columns time, id, type, variable, and value.

pct_error A tibble containing the percentage errors (i.e. error / actual * 100) with the columns time, id, type, variable, and value.

states_train A tibble containing the internal states with the columns time, id, type, variable, and value.

method A list containing several objects and information of the trained ESN (weight matrices, hyperparameters, error and model metrics, etc.).
}
\description{
This function trains an Echo State Network (ESN) to a univariate or multivariate time series.
}
