% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_data.R
\name{trans_data}
\alias{trans_data}
\title{Transform the columns of a numeric matrix}
\usage{
trans_data(data)
}
\arguments{
\item{data}{A numeric matrix containing the values to be transformed. Each column is a variable and each row an observation.}
}
\value{
result A list with the transformed data as numeric matrix and objects of class orderNorm (for inverse transform).
}
\description{
Transform the columns of a numeric matrix using ordered quantile normalizing transformation (OQR)
}
