% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_inputs.R
\name{tune_inputs}
\alias{tune_inputs}
\title{Best subset regression to select model inputs}
\usage{
tune_inputs(
  data,
  lags,
  fourier,
  xreg,
  dy,
  dx,
  n_initial,
  weights,
  scale_inputs,
  inf_crit,
  n_sample,
  n_seed
)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{weights}{Numeric vector. Observation weights for weighted least squares estimation.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("aic", "bic", "hq")}.}

\item{n_sample}{Integer value. The number of random samples for random search.}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}
}
\value{
A \code{list} containing:
  \itemize{
    \item{\code{const}: Logical value. If \code{TRUE}, an intercept term is used.}
    \item{\code{lags}: A \code{list} containing integer vectors with the lags associated with each input variable.}
    \item{\code{n_fourier}: Integer vector. The number of fourier terms (seasonal cycles per period).}
    }
}
\description{
The function \code{tune_inputs()} performs best subset
  regression to identify the optimal model inputs. The models are estimated
  via Ordinary Least Squares (OLS). If the number of all possible models
  exceeds \code{n_sample}, a random search is performed.
}
