% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_pars.R
\name{tune_pars}
\alias{tune_pars}
\title{Estimate information criterion for hyperparameter tuning}
\usage{
tune_pars(
  data,
  pars,
  const,
  lags,
  n_fourier,
  period,
  n_diff,
  n_res,
  n_initial,
  n_seed,
  density,
  inf_crit,
  scale_inputs,
  scale_runif
)
}
\arguments{
\item{data}{A \code{tsibble} containing the time series data.}

\item{pars}{Numeric vector containing the hyperparameters.}

\item{const}{Logical value. If \code{TRUE}, a constant term (intercept) is used.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{n_fourier}{Integer vector. The number of fourier terms (seasonal cycles per period).}

\item{period}{Integer vector. The periodicity of the time series (e.g. \code{period = c(12)} for monthly data or \code{period = c(24, 168)} for hourly data).}

\item{n_diff}{Integer value. The number of differences.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("aic", "bic", "hq")}.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}
}
\value{
model_value Numeric value. The information criterion to be minimized.
}
\description{
The function \code{tune_pars()} estimates the information criterion for hyperparameter
  tuning. The function returns a numeric value (= information criterion), which
  is then minimized within a call to \code{optim()} for varying hyperparameters.
}
