% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_pars.R
\name{tune_pars}
\alias{tune_pars}
\title{Estimate information criterion for hyperparameter tuning}
\usage{
tune_pars(
  data,
  pars,
  inf_crit,
  lags,
  fourier,
  const,
  xreg,
  dy,
  dx,
  n_res,
  n_initial,
  n_seed,
  density,
  type,
  weights,
  penalty,
  scale_runif,
  scale_inputs
)
}
\arguments{
\item{data}{A \code{tsibble} containing the response variable.}

\item{pars}{Numeric vector containing the hyperparameters.}

\item{inf_crit}{Character value. The information criterion \code{inf_crit = c("aic", "bic", "hq")}.}

\item{lags}{A \code{list} containing integer vectors with the lags associated with each input variable.}

\item{fourier}{A \code{list} containing the periods and the number of fourier terms as integer vector.}

\item{const}{Logical value. If \code{TRUE}, an intercept term is used.}

\item{xreg}{A \code{tsibble} containing exogenous variables.}

\item{dy}{Integer vector. The nth-differences of the response variable.}

\item{dx}{Integer vector. The nth-differences of the exogenous variables.}

\item{n_res}{Integer value. The number of internal states within the reservoir (hidden layer).}

\item{n_initial}{Integer value. The number of observations of internal states for initial drop out (throw-off).}

\item{n_seed}{Integer value. The seed for the random number generator (for reproducibility).}

\item{density}{Numeric value. The connectivity of the reservoir weight matrix (dense or sparse).}

\item{type}{Numeric value. The elastic net mixing parameter.}

\item{weights}{Numeric vector. Observation weights for weighted least squares estimation.}

\item{penalty}{Numeric vector. The penalty factors used in \code{glmnet::glmnet()}.}

\item{scale_runif}{Numeric vector. The lower and upper bound of the uniform distribution.}

\item{scale_inputs}{Numeric vector. The lower and upper bound for scaling the time series data.}
}
\value{
model_value Numeric value. The information criterion to be minimized.
}
\description{
The function \code{tune_pars()} estimates the information criterion for hyperparameter
  tuning. The function returns a numeric value (= information criterion), which
  is then minimized within a call to \code{optim()} for varying hyperparameters.
}
